requirements.txt
uvicorn~=0.17.5
fastapi~=0.75.0
loguru~=0.6.0
requests~=2.27.1
redis~=4.1.4
PyExecJS~=1.5.1